build: move forms and models from recognizer to personal account
